Rename to self-growth, fix bugs, update logo, add CI/release workflows v2026.4.8
- Rename bemind-growth to self-growth across crate, package, and DB filename
- Fix f32/f64 type mismatch in ai.rs temperature constants
- Fix Chat test mock setup (list_conversations not returning array)
- Fix pre-existing test failures (Settings version, Goals empty state, Todos
  button selector, AppShell Logo/nav/version, OcrButton error handling)
- Complete Stories list page with expand/collapse and delete
- Update logo to person+growth design with SelfGrowth branding
- Update version to 2026.4.8 (calendar versioning yyyy.m.d)
- Add CI workflow (type check, tests, build, cargo check/clippy)
- Add release workflow (quality gate → build macOS → publish GitHub release)
- Update CLAUDE.md with versioning docs and DB filename

+++ app/src-tauri/src/commands/constants.rs
@@ -11,16 +11,16 @@
 // ---------------------------------------------------------------------------
 
 /// General coaching / chat — warm, conversational.
-pub const TEMP_COACH: f32 = 0.7;
+pub const TEMP_COACH: f64 = 0.7;
 
 /// Analytical insights — lower temperature for structured, data-driven output.
-pub const TEMP_INSIGHTS: f32 = 0.5;
+pub const TEMP_INSIGHTS: f64 = 0.5;
 
 /// Progress summaries — slightly creative but still grounded.
-pub const TEMP_SUMMARIZE: f32 = 0.6;
+pub const TEMP_SUMMARIZE: f64 = 0.6;
 
 /// Chat completion — same balance as coaching.
-pub const TEMP_CHAT: f32 = 0.7;
+pub const TEMP_CHAT: f64 = 0.7;
 
 /// Story generation — higher temperature for creative prose.
 pub const TEMP_STORY: f32 = 0.9;
